+++ CHANGELOG.md
@@ -1,8 +1,8 @@
-# Unreleased
+# v4.2.0 - 2021-10-20
 
 Added
   * AZ-504: Adding diagnostic logging, trust policies & retention options.
-  * AZ-504: Adding Network rules (ip_rules and authorized subnets)
+  * AZ-504: Adding Network rules (`ip_rules` and authorized subnets)
 
 Changed
   * AZ-572: Revamp examples and improve CI